Popover Function
Added a function to interface with the popover in Trello boards, allowing for the card counter and list color menus to display in a native an expected manner.

With this comes a bug introduced when having a completely native popover open, then clicking a trigger for a non-native popover. I may have to find a way to over-ride this or simply leave it be as it doesn't actually break functionality.