Fixed board switching & started percentages
Fixed an issue with count limits not updating when switching boards, it's not an elegant solution but it's one that works.

Started on percentage based colors with a new pair of functions, colors and percentages are to be changed later.

Moved the board ID calculation to a returnable function, making it easier to call.